**TCIO - 123** Implement profile address. 
Also updated additional info logic in figure (TCIO - 136)

**TCIO - 133** Develop reviews logic
- TCIO-144 Develop ReviewEntity and repo. Added ReviewEntity class and updated figure and user entity classes
- TCIO-145  Developed review service layer (created mapper, DTO's, service). Also added interaction between review and other entities (figure, user), so changed figure's (and user's) associated files (such as mapper, DTO's, service).
- TCIO-146 Developed review controller layer logic. Also updated Reviews Service and Security logic in controllers
- TCIO-135 Developed reviews processing. Also overrode equals and hashCode in Figure, User, Review classes

**TCIO - 136, 112, 147, 148** and their subtasks are completed
- TCIO - 136 Additional info
- TCIO - 148 The "[POST] /api/carts" button in "Swagger" returns a 500 error
- TCIO - 147 The "Add figure to wish list" section in "Swagger" returns a 500 error

**TCIO-139**. Updated FigureDTO's (FigureRequestDto.java, FigureResponseDto.java) and FigureMapper.java (added new figure properties for additional info).

**TCIO-112**. Cart logic extended to business state. Rewritten create and update methods. 
Now you can set figure amount in these methods (using Map<String, Integer> where first parameter is figure id and second is figure amount). Also deleted class CartDto.java (it was replaced in previous commit by CartRequestDto.java and CartResponseDto.java).

**TCIO-138** Additional info fields to Figure. 
Updated entity, repo and mapper

**TCIO-117**. Created controller (BannerController.java, with CRUD endpoints) for banner carousel logic. 
Also added new banner bean to Swagger docs (OpenAPIConfig.java).

**TCIO-116** Created services DTO, mapper for Banner.
Changed the names to more abstract ones in entities services methods and mappers

**TCIO-115** Created entity and repository for banner carousel (on main page).

**TCIO-127** implemented GET in stock only logic.
Deleted "Get all figures from wish list" endpoint in WishListController and added Update endpoint to FigureController (also updated FigureMapper)

**TCIO-122**. Updated FigureServiceImpl.java (implemented hot deals logic). 
Also changed pagination logic (instead of getting pagination list (Page<>) from repository I implemented paginateFigureList(from List<> to Page<> with the correct pagination) method and use it when all filtering and sorting parameters are done) to get all figures.

**TCIO-130**. Implemented logic for recently viewed items in services (figure, user). 
Also moved the addFigureToWishList() method from WishListController.java to FigureController.java (so UserServiceImpl.java doesn't need to use FigureServiceImpl.java in its class).

**TCIO-124** (sub TCIO-125, 126, 128). 
Added filtering by price, pagination and get by label (to controller and service) for get all figures method. Also added new CustomNullPointerException.java exception to handle wrong format for start and end price (added its handler method in GlobalExceptionHandler.java).

+++ src/main/java/com/teamChallenge/controller/user/UserController.java
@@ -3,14 +3,16 @@ package com.teamChallenge.controller.user;
 import com.teamChallenge.dto.request.UserRequestDto;
 import com.teamChallenge.dto.request.auth.SignupRequestDto;
 import com.teamChallenge.dto.response.UserResponseDto;
-import com.teamChallenge.entity.user.Roles;
+import com.teamChallenge.entity.user.UserEntity;
 import com.teamChallenge.entity.user.UserServiceImpl;
+import com.teamChallenge.entity.user.address.AddressInfo;
 import com.teamChallenge.exception.CustomErrorResponse;
 import com.teamChallenge.exception.LogEnum;
 import com.teamChallenge.exception.exceptions.generalExceptions.CustomAlreadyExistException;
 import com.teamChallenge.exception.exceptions.generalExceptions.CustomNotFoundException;
 import com.teamChallenge.exception.exceptions.generalExceptions.UnauthorizedAccessException;
 
+import com.teamChallenge.security.AuthUtil;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -44,7 +46,7 @@ public class UserController {
     private static final String URI_USERS_WITH_ID = "/{id}";
 
     private final UserServiceImpl userService;
-
+    private final AuthUtil authUtil;
 
     @PostMapping("/add")
     @Operation(summary = "Add new User")
@@ -58,7 +60,8 @@ public class UserController {
     })
     @SecurityRequirement(name = "BearerAuth")
     public ResponseEntity<UserResponseDto> addUser(@Valid @NotNull @RequestBody SignupRequestDto request, Principal principal) throws CustomAlreadyExistException, UnauthorizedAccessException {
-        validation(principal);
+        authUtil.validateAdminRole(principal);
+
         UserResponseDto user = userService.create(request);
 
         log.info("{}: User (id: {}) has been added", LogEnum.SERVICE, user.id());
@@ -77,7 +80,8 @@ public class UserController {
     })
     @SecurityRequirement(name = "BearerAuth")
     public ResponseEntity<List<UserResponseDto>> userList(Principal principal) throws CustomNotFoundException, UnauthorizedAccessException {
-        validation(principal);
+        authUtil.validateAdminRole(principal);
+
         List<UserResponseDto> users = userService.getAll();
 
         log.info("{}: Users have been retrieved", LogEnum.CONTROLLER);
@@ -97,7 +101,8 @@ public class UserController {
                             schema = @Schema(implementation = CustomErrorResponse.class)) }) })
     @SecurityRequirement(name = "BearerAuth")
     public UserResponseDto getById(@PathVariable String id, Principal principal) throws UnauthorizedAccessException {
-        validation(principal);
+        authUtil.validateAdminRole(principal);
+
         UserResponseDto user = userService.getById(id);
         log.info("{}: User (id: {}) has been retrieved", LogEnum.CONTROLLER, id);
         return user;
@@ -118,9 +123,34 @@ public class UserController {
                             schema = @Schema(implementation = CustomErrorResponse.class)) }) })
     @SecurityRequirement(name = "BearerAuth")
     public UserResponseDto update(@PathVariable String id, @RequestBody UserRequestDto userRequestDto, Principal principal) throws UnauthorizedAccessException {
-        validation(principal);
+        authUtil.validateAdminRole(principal);
+
         UserResponseDto user = userService.update(id, userRequestDto);
         log.info("{}: User (id: {}) has been updated", LogEnum.CONTROLLER, user.id());
+        return user;
+    }
+
+    @PutMapping("/address")
+    @Operation(description = "update user address info")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Updated user address info",
+                    content = {@Content(mediaType = MediaType.APPLICATION_JSON_VALUE,
+                            schema = @Schema(implementation = UserResponseDto.class))}),
+            @ApiResponse(responseCode = "400", description = "Validation error",
+                    content = { @Content(mediaType = "application/json",
+                            schema = @Schema(implementation = CustomErrorResponse.class))}
+            ),
+            @ApiResponse(responseCode = "404", description = "User not found",
+                    content = { @Content(mediaType = "application/json",
+                            schema = @Schema(implementation = CustomErrorResponse.class)) }) })
+    @SecurityRequirement(name = "BearerAuth")
+    public UserResponseDto update(@RequestBody AddressInfo addressInfo, Principal principal) throws UnauthorizedAccessException {
+        //authUtil.validateAdminRole(principal);
+        UserEntity userEntity = userService.findByEmail(principal.getName());
+        userEntity.setAddressInfo(addressInfo);
+
+        UserResponseDto user = userService.update(userEntity);
+        log.info("{}: User (id: {}) address info has been updated", LogEnum.CONTROLLER, user.id());
         return user;
     }
 
@@ -134,15 +164,9 @@ public class UserController {
     @ResponseStatus(HttpStatus.OK)
     @SecurityRequirement(name = "BearerAuth")
     public void delete(@PathVariable String id, Principal principal) throws CustomNotFoundException, UnauthorizedAccessException {
-        validation(principal);
+        authUtil.validateAdminRole(principal);
         userService.delete(id);
 
         log.info("{}: User (id: {}) has been deleted", LogEnum.CONTROLLER, id);
-    }
-
-    private void validation(Principal principal) throws UnauthorizedAccessException {
-        if (!userService.findByEmail(principal.getName()).getRole().equals(Roles.ADMIN)){
-            throw new UnauthorizedAccessException();
-        }
     }
 }
